projects/fred-snyder-examples/airtable-to-markdown/requirements.txt
arrow==0.15.8
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.1
pandas==1.1.0
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wrapt==1.12.1
